+++ src/Whim.Core/Command/ICommandable.cs
@@ -1,4 +1,4 @@
-namespace Whim.Core.Commands
+namespace Whim.Core.Command
 {
     public interface ICommandable
     {
