+++ src/Whim.Core/Monitor/IMonitor.cs
@@ -1,5 +1,7 @@
-namespace Whim.Core {
-    public interface IMonitor {
+namespace Whim.Core.Monitor
+{
+    public interface IMonitor
+    {
         public string Name { get; }
         public int Width { get; }
         public int Height { get; }
