+++ src/Whim.Core/Monitor/IMonitorManager.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
-using Whim.Core.Commands;
 
-namespace Whim.Core
+namespace Whim.Core.Monitor
 {
     public interface IMonitorManager : IEnumerable<IMonitor>, ICommandable
     {
