+++ src/Whim.Core/Window/IWindowLocation.cs
@@ -1,4 +1,4 @@
-namespace Whim.Core
+namespace Whim.Core.Window
 {
     public interface IWindowLocation
     {
